Add files via upload
This was my next or second task under PRODIGY InfoTech to apply Kmeans for clustering the customers using their purchase history for this i used elbow method to find number of clusters can be formed. and we can see by graph that optimal number of cluster can be formed 5 and based on this the kmeans model applied to form clusters of those customers group so the separation and clustering can be seen in graph.